+++ Backend/src/main/java/de/hhn/labfastord/dto/ProductDTO.java
@@ -7,13 +7,13 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductDTO {
-    private Integer productId;
+    private Long productId;
     private String name;
-    private double price;
+    private Double price;
     private String imgName;
     private boolean available;
     private Integer quantity;
-    private Integer categoryId;
+    private Long categoryId;
 
     public ProductDTO() {
     }
